Allow non-root `npm install --global`
This is because the version of npm shipped in versions of nodejs before
v19 can corrupt /usr/bin files. It should never be run as root.

Also downgrade nodejs to v16, which is more commonly usable with node
apps. Stop installing bash-language-server using npm, as a snap exists,
and install the typescript-language-server snap while we're at it.